+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased](https://github.com/craft-ai/craft-ai-client-js/compare/v1.8.0...HEAD) ##
 ### Changed ###
 - Updating the code linter rules to the latest version of [`eslint-config-craft-ai`](https://www.npmjs.com/package/eslint-config-craft-ai).
+- Requests sent to the craft ai have a specific `User-Agent` to identify the used client version & platform.
 
 ### Fixed ###
 - TypeScript type definition of the client method `addAgentContextOperations` now accepts a single context operation.